Fix test exeuction
Refactored project so the tests are in a separate directory to the program
